Import Parsec 1.7.0 source and build with Maven
The Maven build will produce a Parsec.exe compiled with QuickBASIC's BC.EXE and LINK.EXE (executed in DOSBox). This executable, along with the other distributable assets and data files, will be packaged into a ZIP archive for distribution. The contents of this archive match the original Parsec 1.7.0 release as closely as possible.

The DOSBox and QuickBASIC utilities used by the Maven build must be installed on the host machine. The user may set the DOSBOX_HOME and QB_HOME environment variables or specify the dosbox.home and qb.home system properties as Maven command line arguments to tell the build where the tools are installed.