Report clearer error message on failure to read file
Originally, the HandleError subroutine would execute upon failure to read a file. It would print a general file read error message and exit. This was broken during the migration to FreeBASIC, primarily because error handling is not enabled on the compile command line. This breakage resulted in the code trying to read the bad input and then displaying a symbol parsing error, falsely making it seem as though it successfully read content from the input file.

The HandleError subroutine has been retired in favor of simply inserting error checking code in the places where files are being opened for reading. This improves clarity by reducing unnecessary jumping around in the code to follow the flow. Also, the error codes in FreeBASIC are completely different than those in QuickBASIC, so the existing HandleError subroutine no longer behaved as expected anyway.