Locate data.dat file using parsec.exe path
Rather than simply reading data.dat from the current working directory, read it from the same directory containing the parsec.exe. This simplifies the installation procedure (i.e. no need to set working directory in parsec.exe properties) and helps ensure that Parsec will operate correctly regardless of where it gets invoked from.

Another practical benefit of this change is that it enables Parsec to parse a cogscript sinply by drag and drop. That is, drag the .cog file and drop it on parsec.exe to parse it.